refactor: diet date selector as week dates
