+++ src/components/date-selector/DateSelector.tsx
@@ -1,46 +1,49 @@
 import dayjs from "dayjs";
 import { FC } from "react";
 import { AiOutlineCaretLeft, AiOutlineCaretRight } from "react-icons/ai";
+import { useTrackerStore } from "../../store/TrackerStore";
 
-type DateSelectorProps = {
-  selectedDate: string;
-  onLeftArrowClick: () => void;
-  onRightArrowClick: () => void;
-  previousDateDisabled?: boolean;
-  nextDateDisabled?: boolean;
-  showDate?: boolean;
-};
+const today = dayjs().format("YYYY-MM-DD");
 
 const ArrowPlaceholder: FC = () => {
   return <div className="w-[30px] h-[30px]"></div>;
 };
 
-export const DateSelector: FC<DateSelectorProps> = ({
-  selectedDate,
-  onLeftArrowClick,
-  onRightArrowClick,
-  nextDateDisabled,
-  previousDateDisabled,
-  showDate,
-}) => {
+export const DateSelector: FC = () => {
+  const { selectedDate, setSelectedDate } = useTrackerStore(
+    ({ selectedDate, setSelectedDate }) => ({
+      selectedDate,
+      setSelectedDate,
+    })
+  );
+  const setPreviousDate = () => {
+    setSelectedDate(
+      dayjs(selectedDate).subtract(1, "day").format("YYYY-MM-DD")
+    );
+  };
+
+  const setNextDate = () => {
+    setSelectedDate(dayjs(selectedDate).add(1, "day").format("YYYY-MM-DD"));
+  };
+
   return (
     <div className="flex w-full justify-between items-center p-[20px]">
-      {!previousDateDisabled ? (
-        <AiOutlineCaretLeft
-          style={{ fontSize: "30px", cursor: "pointer" }}
-          onClick={onLeftArrowClick}
-        />
-      ) : (
+      {/*  {!previousDateDisabled ? ( */}
+      <AiOutlineCaretLeft
+        style={{ fontSize: "30px", cursor: "pointer" }}
+        onClick={setPreviousDate}
+      />
+      {/* ) : (
         <ArrowPlaceholder />
-      )}
+      )} */}
       <div className="flex flex-col items-center">
         <span>{dayjs(selectedDate).format("dddd")}</span>
-        {showDate && <span>{dayjs(selectedDate).format("DD/MM/YYYY")}</span>}
+        <span>{dayjs(selectedDate).format("DD/MM/YYYY")}</span>
       </div>
-      {!nextDateDisabled ? (
+      {selectedDate !== today ? (
         <AiOutlineCaretRight
           style={{ fontSize: "30px", cursor: "pointer" }}
-          onClick={onRightArrowClick}
+          onClick={setNextDate}
         />
       ) : (
         <ArrowPlaceholder />
