+++ src/pages/TrackerPage.tsx
@@ -4,7 +4,7 @@ import { AiOutlinePlus } from "react-icons/ai";
 import { Accordion } from "../components/Accordion";
 import { ManageFoodModal } from "../components/ManageFoodModal";
 import { Button } from "../components/common/Button";
-import { TrackerDateSelector } from "../components/date-selector/TrackerDateSelector";
+import { DateSelector } from "../components/date-selector/DateSelector";
 import { ModalOverlay } from "../components/ui/ModalOverlay";
 import { useTrackerStore } from "../store/TrackerStore";
 import { Food, IndexedMeals } from "../types";
@@ -170,7 +170,7 @@ function TrackerPage() {
       )}
       <div className="flex flex-col w-full h-full max-h-screen">
         <div className="flex flex-col w-full items-center">
-          <TrackerDateSelector />
+          <DateSelector />
           <span>Summary</span>
           <span>Calories: {totalCalories}</span>
           <div className="flex w-full justify-center gap-x-[10px]">
